+++ test/core.jl
@@ -1215,7 +1215,7 @@ let A = [1]
 end
 
 # Module() constructor
-@test names(Module(:anonymous), all = true, imported = true) == [:anonymous]
+@test names(Module(:anonymous), all = true, imported = true) == [:Base, :anonymous]
 @test names(Module(:anonymous, false), all = true, imported = true) == [:anonymous]
 @test invokelatest(getfield, Module(:anonymous, false, true), :Core) == Core
 @test_throws UndefVarError invokelatest(getfield, Module(:anonymous, false, false), :Core)